Use existing msglevels with hilight level removed
Instead of sending all messages as `MSGLEVEL_PUBLIC`, preserve the
original msglevel of the message and remove the hilight level.

I was in a situation where I both ignored a nick and added it to
`nohilight`:

    /ignore ignored_nick* NO_ACT
    /set nohilight_nicks ignored_nick

With `NO_ACT`, I expect messages from `ignored_nick` not to show
activity on the channel window. However, because all messages passing
through `nohilight` would be posted as `MSGLEVEL_PUBLIC`, the `NO_ACT`
level was ignored.